Fixed userscarried hasnext exception in profile info

+++ Server/netvogue_database/src/main/java/org/netvogue/server/neo4japi/service/UserServiceImpl.java
@@ -105,7 +105,7 @@ public class UserServiceImpl implements UserService{
 		System.out.println(cat);
 		System.out.println(userscarried);
 		return userRepo.doAdvancedSearch(cat, query, userType.toString(), userscarried,
-											fromprice, toprice);
+											fromprice.longValue(), toprice.longValue());
 	}
 	
 	//Network related
